chore(rocm): split correction bias from KDA staging
Keep this branch scoped to caller-owned packed recurrent output. The FP32 correction-bias allocation moves to its standalone accuracy-focused proposal.

Generated-by: OpenAI Codex

+++ vllm/models/kimi_k3/amd/linear.py
@@ -210,10 +210,7 @@ class KimiMoE(nn.Module):
             prefix=f"{prefix}.gate",
         )
 
-        # Match FP32 router logits to avoid a per-forward grouped-top-k cast.
-        self.gate.e_score_correction_bias = nn.Parameter(
-            torch.empty(num_experts, dtype=torch.float32)
-        )
+        self.gate.e_score_correction_bias = nn.Parameter(torch.empty(num_experts))
 
         if self.num_shared_experts is not None:
             shared_intermediate_size = moe_intermediate_size * self.num_shared_experts
